Move to server component and remove unused API route
