--- app/api/properties/search/route.js
@@ -1,43 +1,0 @@
-import connectDB from '@/config/database';
-import Property from '@/models/Property';
-
-// NOTE: here we need to send back a Content-Type: application/json response
-// header rather than a text/plain header.
-
-// GET /api/properties/search
-export const GET = async (request) => {
-  try {
-    await connectDB();
-
-    const { searchParams } = new URL(request.url);
-    const location = searchParams.get('location');
-    const propertyType = searchParams.get('propertyType');
-
-    const locationPattern = new RegExp(location, 'i');
-
-    // Match location pattern against database fields
-    let query = {
-      $or: [
-        { name: locationPattern },
-        { description: locationPattern },
-        { 'location.street': locationPattern },
-        { 'location.city': locationPattern },
-        { 'location.state': locationPattern },
-        { 'location.zipcode': locationPattern },
-      ],
-    };
-
-    // Only check for property if its not 'All'
-    if (propertyType && propertyType !== 'All') {
-      const typePattern = new RegExp(propertyType, 'i');
-      query.type = typePattern;
-    }
-
-    const properties = await Property.find(query);
-
-    return Response.json(properties);
-  } catch (error) {
-    console.log(error);
-    return new Response('Something went wrong', { status: 500 });
-  }
-};
